+++ src/main/java/org/example/server/service/CustomerService.java
@@ -3,5 +3,5 @@ package org.example.server.service;
 import org.example.server.model.entities.Customer;
 
 public interface CustomerService {
-    Customer getCustomerById(Long id);
+    Customer getCustomerById(Integer id);
 }
